+++ Dayana/Server/Api/Controllers/Identity/PermissionController.cs
@@ -4,7 +4,7 @@ using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
 using Dayana.Shared.Basic.MethodsAndObjects.Extension;
 using Dayana.Shared.Persistence.Models.Identity.Filters;
 using Dayana.Shared.Persistence.Models.Identity.Queries.Permissions;
-using Dayana.Shared.Persistence.Models.Identity.Requests.Permissions;
+using Dayana.Shared.Persistence.Models.Identity.Requests;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
