+++ Dayana/Shared/Persistence/Models/Identity/Commands/Auth/LoginCommand.cs
@@ -1,7 +1,5 @@
 ﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
-using Dayana.Shared.Infrastructure.Errors.Identity;
 using Dayana.Shared.Infrastructure.Operations;
-using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Commands.Auth;
