+++ Dayana/Shared/Persistence/Models/Identity/Commands/Roles/CreateRoleCommand.cs
@@ -1,8 +1,5 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants;
-using Dayana.Shared.Basic.MethodsAndObjects.Models;
-using Dayana.Shared.Infrastructure.Errors.Identity;
+﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
 using Dayana.Shared.Infrastructure.Operations;
-using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Commands.Roles;
