+++ Dayana/Shared/Persistence/Models/Identity/Commands/Users/CreateUserCommand.cs
@@ -1,9 +1,6 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants;
-using Dayana.Shared.Basic.MethodsAndObjects.Models;
+﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
 using Dayana.Shared.Domains.Identity.Users;
-using Dayana.Shared.Infrastructure.Errors.Identity;
 using Dayana.Shared.Infrastructure.Operations;
-using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Commands.Users;
