+++ Dayana/Shared/Persistence/Models/Identity/Commands/Users/UpdateUserCommand.cs
@@ -1,8 +1,6 @@
 ﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
 using Dayana.Shared.Domains.Identity.Users;
-using Dayana.Shared.Infrastructure.Errors.Identity;
 using Dayana.Shared.Infrastructure.Operations;
-using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Commands.Users;
