+++ Dayana/Shared/Persistence/Models/Identity/Requests/AuthRequestsFile.cs
@@ -1,5 +1,4 @@
-﻿namespace Dayana.Shared.Persistence.Models.Identity.Requests.Auth;
-
+﻿namespace Dayana.Shared.Persistence.Models.Identity.Requests;
 public class LoginRequest
 {
     public string Username { get; set; }
