+++ Dayana/Shared/Persistence/Models/Identity/Requests/RoleRequestsFile.cs
@@ -1,7 +1,12 @@
 ﻿using Dayana.Shared.Infrastructure.Pagination;
 using Dayana.Shared.Persistence.Models.Identity.Filters;
 
-namespace Dayana.Shared.Persistence.Models.Identity.Requests.Roles;
+namespace Dayana.Shared.Persistence.Models.Identity.Requests;
+public class CreateRoleRequest
+{
+    public string Title { get; set; }
+    public IList<string> PermissionEids { get; set; }
+}
 
 public class GetRolesByFilterRequest : PaginationFilter
 {
@@ -15,4 +20,10 @@ public class GetRolesByFilterRequest : PaginationFilter
     public string? Title { get; set; }
     public List<string>? PermissionEids { get; set; }
     public RoleSortBy? SortBy { get; set; }
+}
+
+public class UpdateRoleRequest
+{
+    public string Title { get; set; }
+    public IList<string> PermissionEids { get; set; }
 }
