--- Dayana/Shared/Persistence/Models/Identity/Requests/Users/CreateUserRequest.cs
@@ -1,12 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Identity.Requests.Users;
-
-public class CreateUserRequest
-{
-    public string FirstName { get; set; }
-    public string LastName { get; set; }
-    public string FullName { get; set; }
-    public string Username { get; set; }
-    public string Mobile { get; set; }
-    public string Password { get; set; }
-    public string Email { get; set; }
-}
