--- Dayana/Shared/Persistence/Models/Identity/Requests/Users/UpdateUserPasswordRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Identity.Requests.Users;
-
-public class UpdateUserPasswordRequest
-{
-    public string NewPassword { get; set; }
-    public string LastPassword { get; set; }
-}
