--- Dayana/Shared/Persistence/Models/Identity/Requests/Users/UpdateUserRolesRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Identity.Requests.Users;
-
-public class UpdateUserRolesRequest
-{
-    public IEnumerable<string> RoleEids { get; set; }
-}
