+++ Dayana/Shared/Persistence/Models/Identity/Validators/CommandValidators/Command_AuthValidatorsFile.cs
@@ -1,11 +1,6 @@
 ﻿using Dayana.Shared.Infrastructure.Errors.Identity;
 using Dayana.Shared.Persistence.Models.Identity.Commands.Auth;
 using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Validators.CommandValidators;
