+++ Dayana/Shared/Persistence/Models/Identity/Validators/QueryValidators/Query_AuthValidatorsFile.cs
@@ -1,13 +1,6 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants;
-using Dayana.Shared.Infrastructure.Errors.Identity;
-using Dayana.Shared.Persistence.Models.Identity.Commands.Roles;
+﻿using Dayana.Shared.Infrastructure.Errors.Identity;
 using Dayana.Shared.Persistence.Models.Identity.Queries.Auth;
 using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Validators.QueryValidators;
 public class GetUserProfileQueryValidator : AbstractValidator<GetUserProfileQuery>
