+++ Dayana/Shared/Persistence/Models/Identity/Validators/QueryValidators/Query_PermissionValidatorsFile.cs
@@ -1,12 +1,6 @@
 ﻿using Dayana.Shared.Infrastructure.Errors.Identity;
-using Dayana.Shared.Persistence.Models.Identity.Queries.Auth;
 using Dayana.Shared.Persistence.Models.Identity.Queries.Permissions;
 using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Validators.QueryValidators;
