+++ Dayana/Shared/Persistence/Models/Identity/Validators/QueryValidators/Query_RoleValidatorsFile.cs
@@ -1,12 +1,6 @@
 ﻿using Dayana.Shared.Infrastructure.Errors.Identity;
-using Dayana.Shared.Persistence.Models.Identity.Queries.Permissions;
 using Dayana.Shared.Persistence.Models.Identity.Queries.Roles;
 using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dayana.Shared.Persistence.Models.Identity.Validators.QueryValidators;
